fix progress bar  .

+++ myexoplayer/src/main/java/com/younchen/myexoplayer/player/DefaultPlayer.java
@@ -106,7 +106,7 @@ public class DefaultPlayer implements Player, ExoPlayer.EventListener {
     }
 
     @Override
-    public float getCurrentPosition() {
+    public long getCurrentPosition() {
         return simpleExoPlayer.getCurrentPosition();
     }
