+++ myexoplayer/src/main/java/com/younchen/myexoplayer/player/Player.java
@@ -23,7 +23,7 @@ public interface Player {
 
     long getDuration();
 
-    float getCurrentPosition();
+    long getCurrentPosition();
 
     long getBufferedPercentage();
